Reintroduce workaround for bug in CGI.pm
Escaped question marks are incorrectly treated as part of the query. This is
bug https://github.com/leejo/CGI.pm/issues/112 which was supposedly fixed in
CGI.pm 4.05, but in fact the fix appears to be incorrect (see
https://github.com/leejo/CGI.pm/issues/112#issuecomment-336141188 ).

Reference the bug report, and put the code that should be used (in
DarkGlass) in a comment.