Pretty DarkGlass URLs.md: simplify the instructions for Apache slightly
There’s no need to capture the URL, as we don’t substitute it in the right
hand side.

Also, don’t ask for instructions for other web servers. This is not a good
place to document the internals of web servers. Keep the Apache
documentation for now. We should perhaps bundle a server with DarkGlass (a
Docker image?) for simple use.