install.sh: fix minor bugs (fix #10, fix #11, fix #13)
Update file we check to ensure we’re being run from the right place (use a
different file, to ensure we’re looking at DarkGlass and not some other
directory with an `About.md`).

Update suffix of `.html` files.

Exit with error if either argument is not a directory, rather than just
giving a warning.