Fix parsing of URL
All these years I have been doing it wrong, in a way that just happened to
work with my Apache setup.

In short, I wasn’t passing any PATH_INFO, so the CGI script had no standard
way to find out which URL was being requested. Instead, it was using the
original URL, which was set by Apache in REQUEST_URI, but this is not a CGI
standard environment variable.

Hence, use `path_info()` to get the path, rather than `url()`, and rewrite
the documented Apache mod_rewrite rule to append the original URI to the CGI
script’s URI, so that PATH_INFO is correctly set when using Apache.